Migrate some tests and fix fallback for empty optional bodies
